minor fixes, removed unecessary comments, formated imports in groups, added/fixed console logs on server side to help with debugging.
