+++ client/src/components/video/Search_bar.js
@@ -24,8 +24,6 @@ class SearchBar extends React.Component{
             </div>
         );
     }
-
 }
    
-
 export default SearchBar;
