+++ client/src/components/video/video_detail.js
